+++ packages/react-email/src/commands/build.ts
@@ -1,13 +1,17 @@
 import fs from 'node:fs';
 import path from 'node:path';
+import { getPackages } from '@manypkg/get-packages';
 import logSymbols from 'log-symbols';
-import { installDependencies, type PackageManagerName, runScript } from 'nypm';
+import { type PackageManagerName, runScript } from 'nypm';
 import ora from 'ora';
 import {
   type EmailsDirectory,
   getEmailsDirectoryMetadata,
 } from '../utils/get-emails-directory-metadata.js';
-import { getPreviewServerLocation } from '../utils/get-preview-server-location.js';
+import {
+  getPreviewServerLocation,
+  isInMonorepo,
+} from '../utils/get-preview-server-location.js';
 import { registerSpinnerAutostopping } from '../utils/register-spinner-autostopping.js';
 
 interface Args {
@@ -17,23 +21,28 @@ interface Args {
 
 const setNextEnvironmentVariablesForBuild = async (
   emailsDirRelativePath: string,
-  builtPreviewAppPath: string,
+  appPath: string,
 ) => {
+  const { rootDir } = await getPackages(appPath);
   const nextConfigContents = `
 import path from 'path';
 const emailsDirRelativePath = path.normalize('${emailsDirRelativePath}');
 const userProjectLocation = '${process.cwd().replace(/\\/g, '/')}';
-const previewServerLocation = '${builtPreviewAppPath.replace(/\\/g, '/')}';
+const previewServerLocation = '${appPath.replace(/\\/g, '/')}';
+const rootDir = ${isInMonorepo ? `'${rootDir.replace(/\\/g, '/')}'` : 'previewServerLocation'};
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   env: {
     NEXT_PUBLIC_IS_BUILDING: 'true',
-    EMAILS_DIR_RELATIVE_PATH: emailsDirRelativePath,
-    EMAILS_DIR_ABSOLUTE_PATH: path.resolve(userProjectLocation, emailsDirRelativePath),
-    PREVIEW_SERVER_LOCATION: previewServerLocation,
-    USER_PROJECT_LOCATION: userProjectLocation
+    REACT_EMAIL_INTERNAL_EMAILS_DIR_RELATIVE_PATH: emailsDirRelativePath,
+    REACT_EMAIL_INTERNAL_EMAILS_DIR_ABSOLUTE_PATH: path.resolve(userProjectLocation, emailsDirRelativePath),
+    REACT_EMAIL_INTERNAL_PREVIEW_SERVER_LOCATION: previewServerLocation,
+    REACT_EMAIL_INTERNAL_USER_PROJECT_LOCATION: userProjectLocation
   },
-  outputFileTracingRoot: previewServerLocation,
+  turbopack: {
+    root: rootDir,
+  },
+  outputFileTracingRoot: rootDir,
   serverExternalPackages: ['esbuild'],
   typescript: {
     ignoreBuildErrors: true
@@ -44,7 +53,7 @@ const nextConfig = {
 export default nextConfig`;
 
   await fs.promises.writeFile(
-    path.resolve(builtPreviewAppPath, './next.config.mjs'),
+    path.resolve(appPath, './next.config.mjs'),
     nextConfigContents,
     'utf8',
   );
@@ -64,7 +73,7 @@ const getEmailSlugsFromEmailDirectory = (
       path
         .join(directoryPathRelativeToEmailsDirectory, filename)
         .split(path.sep)
-        // sometimes it gets empty segments due to trailing slashes
+        // Sometimes it gets empty segments due to trailing slashes
         .filter((segment) => segment.length > 0),
     );
   }
@@ -84,7 +93,7 @@ const getEmailSlugsFromEmailDirectory = (
 // after build
 const forceSSGForEmailPreviews = async (
   emailsDirPath: string,
-  builtPreviewAppPath: string,
+  appPath: string,
 ) => {
   const emailDirectoryMetadata = (await getEmailsDirectoryMetadata(
     emailsDirPath,
@@ -104,15 +113,13 @@ const forceSSGForEmailPreviews = async (
       'utf8',
     );
   };
+  await removeForceDynamic(path.resolve(appPath, './src/app/layout.tsx'));
   await removeForceDynamic(
-    path.resolve(builtPreviewAppPath, './src/app/layout.tsx'),
-  );
-  await removeForceDynamic(
-    path.resolve(builtPreviewAppPath, './src/app/preview/[...slug]/page.tsx'),
+    path.resolve(appPath, './src/app/preview/[...slug]/page.tsx'),
   );
 
   await fs.promises.appendFile(
-    path.resolve(builtPreviewAppPath, './src/app/preview/[...slug]/page.tsx'),
+    path.resolve(appPath, './src/app/preview/[...slug]/page.tsx'),
     `
 
 export function generateStaticParams() { 
@@ -124,8 +131,8 @@ export function generateStaticParams() {
   );
 };
 
-const updatePackageJson = async (builtPreviewAppPath: string) => {
-  const packageJsonPath = path.resolve(builtPreviewAppPath, './package.json');
+const updatePackageJson = async (appPath: string) => {
+  const packageJsonPath = path.resolve(appPath, './package.json');
   const packageJson = JSON.parse(
     await fs.promises.readFile(packageJsonPath, 'utf8'),
   ) as {
@@ -134,22 +141,11 @@ const updatePackageJson = async (builtPreviewAppPath: string) => {
     dependencies: Record<string, string>;
     devDependencies: Record<string, string>;
   };
-  // Turbopack has some errors with the imports in @react-email/tailwind
   packageJson.scripts.build =
     'cross-env NODE_OPTIONS="--experimental-vm-modules --disable-warning=ExperimentalWarning" next build';
   packageJson.scripts.start =
     'cross-env NODE_OPTIONS="--experimental-vm-modules --disable-warning=ExperimentalWarning" next start';
-  delete packageJson.scripts.postbuild;
 
-  packageJson.name = 'preview-server';
-
-  for (const [dependency, version] of Object.entries(
-    packageJson.devDependencies,
-  )) {
-    packageJson.devDependencies[dependency] = version.replace('workspace:', '');
-  }
-
-  delete packageJson.devDependencies['@react-email/components'];
   delete packageJson.scripts.prepare;
 
   await fs.promises.writeFile(
@@ -174,6 +170,10 @@ export const build = async ({
 
     spinner.text = `Checking if ${emailsDirRelativePath} folder exists`;
     if (!fs.existsSync(emailsDirRelativePath)) {
+      spinner.stopAndPersist({
+        symbol: logSymbols.error,
+        text: `Directory does not exist: ${emailsDirRelativePath}`,
+      });
       process.exit(1);
     }
 
@@ -187,19 +187,31 @@ export const build = async ({
       await fs.promises.rm(builtPreviewAppPath, { recursive: true });
     }
 
-    spinner.text = 'Copying preview app from CLI to `.react-email`';
+    spinner.text = 'Copying UI source to `.react-email`';
     await fs.promises.cp(previewServerLocation, builtPreviewAppPath, {
       recursive: true,
       filter: (source: string) => {
-        // do not copy the CLI files
+        const relativeSource = path.relative(previewServerLocation, source);
         return (
-          !/(\/|\\)cli(\/|\\)?/.test(source) &&
-          !/(\/|\\)\.next(\/|\\)?/.test(source) &&
-          !/(\/|\\)\.turbo(\/|\\)?/.test(source) &&
-          !/(\/|\\)node_modules(\/|\\)?$/.test(source)
+          !/\.next/.test(relativeSource) &&
+          !/\.turbo/.test(relativeSource) &&
+          !/node_modules\/.bin/.test(relativeSource)
         );
       },
     });
+
+    await fs.promises.cp(
+      path.resolve(previewServerLocation, 'node_modules/.bin'),
+      path.resolve(builtPreviewAppPath, 'node_modules/.bin'),
+      {
+        recursive: true,
+        // With this enabled, means the symlinks remain relative, which is
+        // what we want for copying it over. If we don't enable this,
+        // it resolves the symlink to the original location on disk,
+        // which ends up causing unexpected errors.
+        verbatimSymlinks: true,
+      },
+    );
 
     if (fs.existsSync(staticPath)) {
       spinner.text =
@@ -225,13 +237,6 @@ export const build = async ({
 
     spinner.text = "Updating package.json's build and start scripts";
     await updatePackageJson(builtPreviewAppPath);
-
-    spinner.text = 'Installing dependencies on `.react-email`';
-    await installDependencies({
-      cwd: builtPreviewAppPath,
-      silent: true,
-      packageManager,
-    });
 
     spinner.stopAndPersist({
       text: 'Successfully prepared `.react-email` for `next build`',